Enhance TickerServer and refactor other classes
In TickerServer, functionality is extended to include a stockFeed list and a symbolToDisplayName map for symbol and company name mapping. Also, a mock data faker is used for generating company names and stock feeds. In addition, add typecasting in MockDataProducer and revise keys to be "even" and "odd" in KTableCountExample.